test is not qt specific
