+++ traitsui/tests/test_ui_panel.py
@@ -14,7 +14,8 @@
 import unittest
 
 from traits.api import HasTraits, Int
-from traitsui.api import HGroup, Item, spring, VGroup, View
+from traitsui.api import HelpButton, HGroup, Item, spring, VGroup, View
+from traitsui.testing.api import MouseClick, UITester
 from traitsui.tests._tools import (
     BaseTestMixin,
     create_ui,
@@ -27,6 +28,18 @@ class ObjectWithNumber(HasTraits):
     number1 = Int()
     number2 = Int()
     number3 = Int()
+
+
+class HelpPanel(HasTraits):
+    my_int = Int(2, help='this is the help for my int')
+
+    def default_traits_view(self):
+        view = View(
+            Item(name="my_int"),
+            title="HelpPanel",
+            buttons=[HelpButton],
+        )
+        return view
 
 
 @requires_toolkit([ToolkitName.qt, ToolkitName.wx])
@@ -55,3 +68,11 @@ class TestUIPanel(BaseTestMixin, unittest.TestCase):
         # This should not fail.
         with create_ui(obj1, dict(view=view)):
             pass
+
+    # Regression test for enthought/traitsui#1538
+    def test_show_help(self):
+        panel = HelpPanel()
+        tester = UITester()
+        with tester.create_ui(panel) as ui:
+            help_button = tester.find_by_id(ui, 'Help')
+            help_button.perform(MouseClick())
